+++ src/api/user.ts
@@ -1,6 +1,6 @@
 import { fetcher } from '@/api/fetcher';
-import { BaseResponse } from '@/types/base';
-import { User } from '@/types/User';
+import { BaseResponse } from '@/types/common/base';
+import { User } from '@/types/user/user';
 
 const getUserInfo = async (): Promise<BaseResponse<User>> => {
   return await fetcher<BaseResponse<User>>('/v1/users/info');
