+++ src/app/(withHeader)/layout.tsx
@@ -1,6 +1,6 @@
 import { ReactNode } from 'react';
-import Header from '@/components/Header';
-import Sidebar from '@/components/Sidebar';
+import Header from '@/components/common/Header';
+import Sidebar from '@/components/common/Sidebar';
 
 export default function WithHeaderLayout({
                                            children,
