+++ src/app/(withHeader)/login/page.tsx
@@ -1,6 +1,6 @@
 "use client"
 
-import Button from '@/components/Button';
+import Button from '@/components/common/Button';
 import Image from 'next/image';
 import kakaoLogin from '@/assets/kakao-login.png'
 import logo from '@/assets/logo.png'
