+++ src/components/chatbot/BotProfile.tsx
