+++ src/components/post/NewsListItem.tsx
@@ -1,4 +1,4 @@
-import { News } from '@/features/Post/types/News';
+import { News } from '@/types/post/news';
 import { ChevronRight } from 'lucide-react';
 import { differenceInHours, parse } from 'date-fns';
 import { formatTime } from '@/utils/formatTime';
