+++ src/components/post/PostHeader.tsx
@@ -1,4 +1,4 @@
-import { PostPlatform } from '@/features/Post/types/PostPlatform';
+import { PostPlatform } from '@/types/post/postPlatform';
 
 interface PostHeaderProps {
   tag?: string | null;
