+++ src/types/common/base.ts
