+++ src/types/common/headline.ts
