+++ src/types/post/newsDetail.ts
